// src/ai/types/unhelpfulAnswerDescription.ts
export enum UnhelpfulAnswerDescription {
  OFF_TOPIC = 'offTopic',
  INCORRECTLY_UNDERSTOOD_REQUEST = 'incorrectlyUnderstoodRequest',
  TOO_GENERAL = 'tooGeneral',
  NOT_ENOUGH_SPECIFIC_ACTIONABLE_STEPS = 'notEnoughSpecificsActionableSteps',
  INACCURATE_QUESTIONABLE_FACTS = 'inaccurateQuestionableFacts',
  TOO_LONG = 'tooLong',
  WRONG_TONE_TOO_DRY_TOO_INTRUSIVE = 'wrongToneTooDryTooIntrusive',
  OTHER = 'other',
}